refactored to use View and Model
