+++ js/accordionModel.js
@@ -1,0 +1,11 @@
+define([
+    'core/js/models/itemsModel'
+], function(ItemsModel) {
+
+    var AccordionModel = ItemsModel.extend({
+
+    });
+
+    return AccordionModel;
+
+});
